requirements.txt
Streamlit
google-generativeai
python-dotenv  
gtts 
streamlit-option-menu